requirements.txt
tavily-python
requests
pytest
pytest-asyncio
beautifulsoup4
spacy
python-dotenv
